use of dataset values in deletion

+++ front/js/cart.js
@@ -27,10 +27,12 @@ cart.map(
       .then(api => {
 
         cartTemplate(api, element)
-        articlesSum(api, element)
+
+
+
         modifQuantity(api, element)
         deleteItems(api, element)
-
+        articlesSum(api, element)
       })
 
       .catch(error => {
@@ -64,7 +66,7 @@ function cartTemplate(api, element) {
                       <input type="number" class="itemQuantity" name="itemQuantity" min="1" max="100" value=${element.quantities} >
                     </div>
                     <div class="cart__item__content__settings__delete">
-                      <p class="deleteItem" >Supprimer</p>
+                      <p  data-id = "${element.id}" data-color= "${element.colours}" class="deleteItem" >Supprimer</p>
 
                     </div>
                   </div>
@@ -82,6 +84,8 @@ function cartTemplate(api, element) {
  * display the sum of prices 
  */
 //-----------------------------------------------------------------------------
+document.getElementById('totalQuantity').textContent = 0
+document.getElementById('totalPrice').textContent = totalPrix = 0
 
 function articlesSum() {
   let totalPrix = 0;
@@ -115,10 +119,10 @@ function deleteItems(api, element) {
 
   Array.from(supprime).forEach(function (deleteButton) {
 
-    deleteButton.addEventListener("click", function () {
+    deleteButton.addEventListener("click", function (e) {
 
-      deleteButton.closest('article').remove()
-      window.localStorage.removeItem(element.id + element.colours)
+      deleteButton.closest('article').remove();
+      window.localStorage.removeItem(e.target.dataset.id + e.target.dataset.color);
 
       articlesSum(api, element);
 
@@ -150,7 +154,7 @@ function modifQuantity(api, element) {
       else {
 
         for (article of cart) {
-          article.quantities = e.target.value
+          article.quantities = Number(e.target.value)
           let product = article.id + article.colours
           window.localStorage.setItem(product, JSON.stringify(article))
           te.dataset.quantité = e.target.value;
@@ -177,9 +181,9 @@ function getForm() {
   let form = document.querySelector(".cart__order__form");
 
   let emailRegExp = new RegExp('^[a-zA-Z0-9.-_]+[@]{1}[a-zA-Z0-9.-_]+[.]{1}[a-z]{2,10}$');
-  let charRegExp = new RegExp("^[a-zA-Z ,.'-]+$");
-  let addressRegExp = new RegExp("^[0-9]{1,3}(?:(?:[,. ]){1}[-a-zA-Zàâäéèêëïîôöùûüç]+)+",);
-
+  let charRegExp = new RegExp("[a-zA-Z ,.'-]+$");
+  //let addressRegExp = new RegExp("^[0-9](?:(?:[,. ]){1}[-a-zA-Zàâäéèêëïîôöùûüç]+)+",);
+  let addressRegExp = /^[^±!@£$%^&*_+¡€#¢§¶•ªº()"«\\/\{\}\[\]\~<>?:;|=.]+$/;
   form.firstName.addEventListener('change', function () {
     validFirstName(this);
   });
@@ -205,41 +209,60 @@ function getForm() {
   const validFirstName = function (inputFirstName) {
     let firstNameErrorMsg = inputFirstName.nextElementSibling;
 
-    if (charRegExp.test(inputFirstName.value) == false) {
-
-      firstNameErrorMsg.innerHTML = 'Veuillez renseigner ce champ.';
-      form.firstName.style.border = "solid 1.5px red"
-
+    if (charRegExp.test(inputFirstName.value) == true) {
+      form.firstName.style.border = "none"
+      firstNameErrorMsg.innerHTML = '';
     }
+    else {
+
+      firstNameErrorMsg.innerHTML = 'Veuillez renseigner votre prénom.';
+      form.firstName.style.border = "solid 1.5px red"
+    }
+
   };
 
   const validLastName = function (inputLastName) {
     let lastNameErrorMsg = inputLastName.nextElementSibling;
 
-    if (charRegExp.test(inputLastName.value) == false) {
-
-      lastNameErrorMsg.innerHTML = 'Veuillez renseigner ce champ.';
-      form.lastName.style.border = "solid 1.5px red"
-
+    if (charRegExp.test(inputLastName.value) == true) {
+      form.lastName.style.border = "none"
+      lastNameErrorMsg.innerHTML = '';
     }
-  };
+
+    else {
+      lastNameErrorMsg.innerHTML = 'Veuillez renseigner votre nom.';
+      form.lastName.style.border = "solid 1.5px red"
+    }
+
+
+  }
+
 
   const validAddress = function (inputAddress) {
     let addressErrorMsg = inputAddress.nextElementSibling;
 
-    if (addressRegExp.test(inputAddress.value) == false) {
+    if (addressRegExp.test(inputAddress.value) == true) {
+      addressErrorMsg.innerHTML = '';
+      form.address.style.border = "none"
+    }
 
-      addressErrorMsg.innerHTML = 'Veuillez renseigner ce champ.';
-      form.address.style.border = "solid 1.5px red"
+    else {
+      addressErrorMsg.innerHTML = 'Veuillez renseigner une adresse valide.';
+      form.address.style.border = "solid 1.5px red";
     }
   };
 
   const validCity = function (inputCity) {
     let cityErrorMsg = inputCity.nextElementSibling;
 
-    if (charRegExp.test(inputCity.value) === false) {
+    if (charRegExp.test(inputCity.value) == true) {
+      cityErrorMsg.innerHTML = '';
+      form.city.style.border = "none"
+    }
 
-      cityErrorMsg.innerHTML = 'Veuillez renseigner ce champ.';
+    else {
+
+      cityErrorMsg.innerHTML = 'Veuillez renseigner votre ville.';
       form.city.style.border = "solid 1.5px red"
     }
   };
@@ -247,9 +270,12 @@ function getForm() {
   const validEmail = function (inputEmail) {
     let emailErrorMsg = inputEmail.nextElementSibling;
 
-    if (emailRegExp.test(inputEmail.value) === false) {
-
-      emailErrorMsg.innerHTML = 'Veuillez renseigner votre email.';
+    if (emailRegExp.test(inputEmail.value) == true) {
+      form.email.style.border = "none"
+      emailErrorMsg.innerHTML = '';
+    }
+    else {
+      emailErrorMsg.innerHTML = 'Veuillez renseigner une adress email valide.';
       form.email.style.border = "solid 1.5px red"
     }
   };
@@ -273,16 +299,11 @@ getForm();
 
 
 function postForm() {
-  const btn_commander = document.getElementById("order");
+  const orderButton = document.getElementById("order");
+  let form = document.querySelector(".cart__order__form");
 
 
-
-
-  btn_commander.addEventListener("click", (event) => {
-
-
-
-
+  orderButton.addEventListener("click", (e) => {
 
     let inputName = document.getElementById('firstName');
     let inputLastName = document.getElementById('lastName');
@@ -290,50 +311,68 @@ function postForm() {
     let inputCity = document.getElementById('city');
     let inputMail = document.getElementById('email');
 
-
-
-    let productId = []
-
-    for (let i = 0; i < localStorage.length; i++) {
-      let productParse = JSON.parse(window.localStorage.getItem(localStorage.key(i))).id
-      productId.push(productParse)
-
+    if (localStorage.length == 0) {
+      e.preventDefault();
+      alert('Veuillez choisir un article avant de passez votre commande.')
+    }
+    else if (
+      inputAdress.value === '' ||
+      inputCity.value === '' ||
+      inputLastName.value === '' ||
+      inputMail.value === '' ||
+      inputName.value === ''
+    ) {
+      e.preventDefault();
+      alert('Veuillez renseigner tout les champs.')
     }
 
-    const order = {
-      contact: {
-        firstName: inputName.value,
-        lastName: inputLastName.value,
-        address: inputAdress.value,
-        city: inputCity.value,
-        email: inputMail.value,
-      },
-      products: productId,
+    else {
+
+      let productId = []
+
+      for (let i = 0; i < localStorage.length; i++) {
+        let productParse = JSON.parse(window.localStorage.getItem(localStorage.key(i))).id
+        productId.push(productParse)
+
+      }
+
+      const order = {
+        contact: {
+          firstName: inputName.value,
+          lastName: inputLastName.value,
+          address: inputAdress.value,
+          city: inputCity.value,
+          email: inputMail.value,
+        },
+        products: productId,
+      }
+
+
+      fetch("http://localhost:3000/api/products/order", {
+
+        method: 'POST',
+        body: JSON.stringify(order),
+        headers: {
+          'Accept': 'application/json',
+          "Content-Type": "application/json"
+        },
+      })
+
+        .then((response) => response.json())
+        .then((data) => {
+
+          localStorage.clear();
+          localStorage.setItem("orderId", data.orderId);
+          document.location.href = "confirmation.html";
+        })
+        .catch(error => {
+          alert(" Erreur : " + error.message);
+        })
+        ;
+
     }
-
-
-    fetch("http://localhost:3000/api/products/order", {
-
-      method: 'POST',
-      body: JSON.stringify(order),
-      headers: {
-        'Accept': 'application/json',
-        "Content-Type": "application/json"
-      },
-    })
-
-      .then((response) => response.json())
-      .then((data) => {
-
-        localStorage.clear();
-        localStorage.setItem("orderId", data.orderId);
-        document.location.href = "confirmation.html";
-      })
-      .catch(error => {
-        alert(" Erreur : " + error.message);
-      })
-      ;
   })
+
 }
 postForm();
